requirements.txt
pyqt5
pyqt5-tools
numpy
Pillow